[refactor] renaming again, new file storage.rs
