+++ src/net/behavior.rs
