+++ src/net/storage.rs
@@ -1,0 +1,1 @@
+//! wraps up "net storage" which is implemented by the kademlia functionality.
